Обновление классов и добавление новых функциональностей
Изменены модификаторы доступа на `public` для классов `ClothingSize`, `Item`, `ItemType`, `Name`, `Sku` и `StockItem`.
Добавлены свойства `Quantity` и `MinimalQuantity` в `StockItem`, а также методы для управления количеством товаров и установки размера одежды.
Введены новые классы `Quantity` и `Tag` как объекты значений.
Добавлены интерфейсы `IStockItemRepository` и `IUnitOfWork` для работы с репозиториями.
Реализованы команды `CreateStockItemCommand` и `GiveOutStockItemCommand` с соответствующими обработчиками.
Добавлены события домена для обработки минимального количества товаров.
Обновлены проектные файлы и тесты для проверки новых функциональностей.

+++ OzonEdu.StockApi.Domain/Models/Entity.cs
@@ -11,7 +11,7 @@ namespace OzonEdu.StockApi.Domain.Models
 	public abstract class Entity
 	{
 		protected int? _requestedHashCode;
-		public virtual long Id { get; protected set; }
+		public virtual int Id { get; protected set; }
 		private readonly List<INotification> _domainEvents = new();
 		public IReadOnlyCollection<INotification> DomainEvents => _domainEvents.AsReadOnly();
 		public void AddDomainEvent(INotification eventItem)
